requirements.txt
streamlit>=1.38.0
openai>=1.40.0
pydantic>=2.0.0
graphviz>=0.20.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pytest>=8.0.0
